+++ source/modules/dgl/src/ImageKnob.cpp
@@ -224,8 +224,8 @@ void ImageKnob::onDisplay()
         const GLint w2 = getWidth()/2;
         const GLint h2 = getHeight()/2;
 
-        glTranslatef(getX()+w2, getY()+h2, 0.0f);
-        glRotatef(normValue*fRotationAngle, 0.0f, 0.0f, 1.0f);
+        glTranslatef(static_cast<float>(getX()+w2), static_cast<float>(getY()+h2), 0.0f);
+        glRotatef(normValue*static_cast<float>(fRotationAngle), 0.0f, 0.0f, 1.0f);
 
         glBegin(GL_QUADS);
           glTexCoord2f(0.0f, 0.0f);
